requirements.txt
Flask
groq
sqlalchemy
python-dotenv
pytz
requests